Adjust listings package use/advice (#14)
* remove most lstlisting syntax highlighting

Remove the language=xxx optional syntax highlighting parameter from
most of the uses of the lstlisting environment in the source,
on the grounds that the output is usually more distracting than helpful;
at least IMHO and as discussed and not disagreed with at issue #8.

* Adjust pretty-printing advice in verbatim section

Note that language-sensitive formatting in listings environments
is not always perfect, and advise authors to think about whether
to use it or not.  Also add a sentence explaining why listings
can provide more value than vanilla verbatim.